requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py37h5e8e339_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py37h5e8e339_1001
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.5.30=ha878542_0
cachetools=4.2.2=pyhd8ed1ab_0
cairo=1.14.12=he6fea26_5
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.6=py37hc58025e_0
chardet=4.0.0=py37h89c1867_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=py37h89c1867_0
cryptography=3.4.7=py37h5d9358c_0
cudatoolkit=10.2.89=hfd86e86_1
dataclasses=0.8=pyhc8e2a94_1
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=he4413a7_1000
freeglut=3.2.1=h9c3ff4c_2
freetype=2.10.4=h5ab3b9f_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.55.0=0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.1=pyh6c4a22f_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.13=h58526e2_1001
grpcio=1.38.1=py37hb27c1af_0
harfbuzz=1.9.0=h08d66d9_0
hdf5=1.10.2=hc401514_3
icu=58.2=hf484d3e_1000
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.6.1=py37h89c1867_0
intel-openmp=2021.3.0=h06a4308_3350
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran=3.0.0=1
libglu=9.0.0=he1b5a44_1001
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h780b84a_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.9=h13577e0_2
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pyhd8ed1ab_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
multidict=5.1.0=py37h5e8e339_1
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h7f98852_0
pcre=8.45=h9c3ff4c_0
pillow=8.3.1=py37h2c7a002_0
pip=21.1.3=py37h06a4308_0
pixman=0.34.0=h14c3975_1003
protobuf=3.17.2=py37hcd2ae1e_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.1.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py37h89c1867_3
python=3.7.10=h12debd9_4
python_abi=3.7=2_cp37m
pytorch=1.9.0=py3.7_cuda10.2_cudnn7.6.5_0
pyu2f=0.1.5=pyhd8ed1ab_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
setuptools=52.0.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.5.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py37h7f0c10b_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py37
torchvision=0.2.1=py37_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py37h5e8e339_2
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
